pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "watchdog>=6.0.0",
]
